ETLWeather/requirements.txt
apache-airflow-providers-http
apache-airflow-providers-postgres
requests
pandas
